test(tuika): split monolithic tests.rs into per-module suites
Move the ~1.9k-line tests.rs into `#[cfg(test)] mod tests` blocks colocated
with the code they exercise, matching the convention already used by
width.rs, runner.rs, hyperlink.rs, and clipboard.rs. Shared helpers
(buffer, row, rainbow_theme, render_el, render_view_rows) move into a
cfg(test)-only test_support module so they are defined once. The remaining
genuinely cross-cutting checks stay in the renamed integration.rs.

No test logic changes: the full set of test names is identical before and
after (135 unit tests), and clippy/fmt are clean.
