+++ crates/tuika/src/components/select.rs
@@ -252,3 +252,139 @@ impl SelectList {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::event::{Event, EventFlow, Key, KeyCode};
+    use crate::style::Theme;
+    use crate::test_support::{buffer, rainbow_theme, row};
+    use crate::view::{RenderCtx, View};
+    use crate::{Size, Surface};
+    use ratatui::text::Line;
+
+    #[test]
+    fn select_navigation_wraps_and_confirms() {
+        let mut s = SelectState::new();
+        let down = Event::Key(Key::new(KeyCode::Down));
+        let up = Event::Key(Key::new(KeyCode::Up));
+        assert_eq!(s.handle(&up, 3), SelectOutcome::Moved(EventFlow::Consumed));
+        assert_eq!(s.selected(), 2); // wrapped from 0 to last
+        assert_eq!(
+            s.handle(&down, 3),
+            SelectOutcome::Moved(EventFlow::Consumed)
+        );
+        assert_eq!(s.selected(), 0); // wrapped back
+        let enter = Event::Key(Key::new(KeyCode::Enter));
+        assert_eq!(s.handle(&enter, 3), SelectOutcome::Confirmed(0));
+        let esc = Event::Key(Key::new(KeyCode::Esc));
+        assert_eq!(s.handle(&esc, 3), SelectOutcome::Cancelled);
+    }
+
+    #[test]
+    fn select_move_up_down_clamp_at_ends() {
+        let mut s = SelectState::new();
+        // Down steps forward, clamping at the last of `len` rows (no wrap).
+        s.move_down(3);
+        assert_eq!(s.selected(), 1);
+        s.move_down(3);
+        assert_eq!(s.selected(), 2);
+        s.move_down(3);
+        assert_eq!(s.selected(), 2); // held at the bottom, not wrapped
+        // Up steps back, clamping at the top.
+        s.move_up();
+        assert_eq!(s.selected(), 1);
+        s.move_up();
+        s.move_up();
+        assert_eq!(s.selected(), 0); // held at the top, not wrapped
+        // Degenerate: an empty list stays at 0.
+        s.move_down(0);
+        assert_eq!(s.selected(), 0);
+    }
+
+    #[test]
+    fn select_state_select_sets_index_directly() {
+        // A host can drive the highlight from its own state.
+        let mut s = SelectState::new();
+        s.select(2);
+        assert_eq!(s.selected(), 2);
+    }
+
+    #[test]
+    fn select_highlights_current_row() {
+        let items = vec![Line::from("alpha"), Line::from("beta")];
+        let mut state = SelectState::new();
+        state.handle(&Event::Key(Key::new(KeyCode::Down)), 2); // select beta
+        let list = SelectList::new(items, &state);
+        let mut buf = buffer(10, 2);
+        let theme = Theme::default();
+        let ctx = RenderCtx::new(&theme);
+        let area = buf.area;
+        let mut surface = Surface::new(&mut buf, area);
+        list.render(area, &mut surface, &ctx);
+        assert!(row(&buf, 1).contains("beta"));
+        // Selected row carries the selection background.
+        assert_eq!(buf[(0, 1)].bg, theme.selection_bg);
+        assert_eq!(buf[(0, 0)].bg, ratatui::style::Color::Reset);
+    }
+
+    #[test]
+    fn select_viewport_windows_a_long_list_and_keeps_selection_visible() {
+        // 20 items, viewport of 4: the selection must always be on screen.
+        let items: Vec<Line> = (0..20).map(|i| Line::from(format!("item{i}"))).collect();
+        let mut state = SelectState::new();
+        state.select(12);
+        let theme = Theme::default();
+        let list = SelectList::new(items.clone(), &state).viewport(4);
+        // Windowed height is the viewport, not the full list.
+        assert_eq!(list.measure(Size::new(20, 40)).height, 4);
+        let rendered = crate::testing::render(&list, 20, 4, &theme);
+        let text = crate::testing::grid(&rendered);
+        assert!(
+            text.contains("item12"),
+            "selection should be visible:\n{text}"
+        );
+        assert!(
+            !text.contains("item0\n") && !text.contains("item19"),
+            "far items windowed out"
+        );
+        // A scrollbar occupies the last column on at least one row.
+        let has_scrollbar = (0..4).any(|y| matches!(rendered[(19, y)].symbol(), "█" | "│"));
+        assert!(
+            has_scrollbar,
+            "overflowing list should draw a scrollbar:\n{text}"
+        );
+    }
+
+    #[test]
+    fn select_viewport_shows_whole_list_when_it_fits() {
+        let items: Vec<Line> = (0..3).map(|i| Line::from(format!("item{i}"))).collect();
+        let state = SelectState::new();
+        let list = SelectList::new(items, &state).viewport(8);
+        // Fits within the viewport → no windowing, height is the item count.
+        assert_eq!(list.measure(Size::new(20, 40)).height, 3);
+        let theme = Theme::default();
+        let text = crate::testing::grid(&crate::testing::render(&list, 20, 3, &theme));
+        assert!(text.contains("item0") && text.contains("item2"));
+    }
+
+    #[test]
+    fn select_list_selection_uses_theme_slots() {
+        let t = rainbow_theme();
+        let mut state = SelectState::new();
+        state.handle(&Event::Key(Key::new(KeyCode::Down)), 2); // select row 1
+        let list = SelectList::new(vec![Line::from("a"), Line::from("b")], &state);
+        let mut buf = buffer(10, 2);
+        let area = buf.area;
+        let ctx = RenderCtx::new(&t);
+        let mut surface = Surface::new(&mut buf, area);
+        list.render(area, &mut surface, &ctx);
+        assert_eq!(buf[(0, 1)].bg, t.selection_bg, "selected row bg");
+        assert_eq!(buf[(0, 1)].fg, t.selection_fg, "selected caret fg");
+        assert_ne!(
+            buf[(0, 0)].bg,
+            t.selection_bg,
+            "unselected row not highlighted"
+        );
+    }
+}
